Use new object in callback
